requirements.txt
# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# AI/ML Pipeline
ultralytics>=8.0.0
deep-sort-realtime>=1.3.2
opencv-python>=4.8.0
numpy>=1.24.0

# Video Processing
ffmpeg-python>=0.2.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0

# Face Recognition (primary: InsightFace ArcFace, fallback: facenet-pytorch)
insightface>=0.7.3
onnxruntime>=1.16.0          # CPU inference for InsightFace models
# onnxruntime-gpu>=1.16.0    # Uncomment instead of above if you have CUDA 11.x+
facenet-pytorch>=2.5.3       # fallback if InsightFace is unavailable
